move functions folder to root project
