+++ dags/etl/dbt_trigger.py
@@ -7,7 +7,7 @@ from airflow.decorators import dag
 from airflow.operators.python import PythonOperator
 from cosmos import DbtTaskGroup, ProjectConfig, RenderConfig
 
-from dags.functions.alert_email import on_failure_callback, send_status_email
+from functions.alert_email import on_failure_callback, send_status_email
 from include.constants import jaffle_shop_path, venv_execution_config
 from include.profiles import redshift_db
