+++ dags/etl/elt_criptoya_other.py
@@ -6,10 +6,10 @@ from types import NoneType
 from airflow.decorators import dag
 from airflow.operators.python import PythonOperator
 
-from dags.functions.alert_email import on_failure_callback, send_status_email
-from dags.functions.extract_data import extract_data_from_api
-from dags.functions.load_data import load_data_to_redshift
-from dags.functions.transform_data import transform_other_usd_from_criptoya_api
+from functions.alert_email import on_failure_callback, send_status_email
+from functions.extract_data import extract_data_from_api
+from functions.load_data import load_data_to_redshift
+from functions.transform_data import transform_other_usd_from_criptoya_api
 
 ETL_NAME = "CriptoYa Other"
