+++ tests/functions/test_extract_function.py
@@ -3,7 +3,7 @@ from unittest.mock import patch
 import pytest
 from requests.exceptions import RequestException
 
-from dags.functions.extract_data import extract_data_from_api
+from functions.extract_data import extract_data_from_api
 
 
 @pytest.fixture
