+++ tests/functions/test_load_function.py
@@ -1,12 +1,13 @@
+from unittest.mock import patch
+
 import pytest
-import pandas as pd
-from unittest.mock import patch, MagicMock
-from dags.functions.load_data import connect_to_redshift_engine, load_data_to_redshift
+
+from functions.load_data import load_data_to_redshift
 
 
 @pytest.fixture
 def mock_create_engine():
-    with patch("dags.functions.load_data.create_engine") as mock:
+    with patch("functions.load_data.create_engine") as mock:
         yield mock
 
 
@@ -29,7 +30,6 @@ def test_connect_to_redshift_engine(mock_create_engine):
     Test connect_to_redshift_engine function to ensure it creates an engine.
     """
     connection_string = "redshift+psycopg2://2024_diego_rojas:password@redshift-pda-cluster.cnuimntownzt.us-east-2.redshift.amazonaws.com:5439/pda"
-    engine = connect_to_redshift_engine()
 
     mock_create_engine.assert_called_once()
 
@@ -43,7 +43,7 @@ def test_load_data_to_redshift(mock_create_engine, sample_json_data):
     """
     mock_engine = mock_create_engine.return_value
 
-    with patch("dags.functions.load_data.pd.DataFrame.to_sql") as mock_to_sql:
+    with patch("functions.load_data.pd.DataFrame.to_sql") as mock_to_sql:
         load_data_to_redshift(sample_json_data, "mock_table")
 
         mock_to_sql.assert_called_once_with(
